Add leaderboard history top 100 player from every league
From this season and last season. This should help with getting a better local database to improve the bots playing strategy and also in including cards that I get or rent for the bot to use since bigger database means that the possibility of the bot seeing plays with those cards used increases so that the chances of the bot using the cards increases. Combinate this with combining json histories of different days (pull and combine into a bigger json file) meaning a db file